[Refactor][Connector] Refactor connector interface
