+++ datavines-connector/datavines-connector-api/src/main/java/io/datavines/connector/api/ConfigBuilder.java
@@ -1,0 +1,6 @@
+package io.datavines.connector.api;
+
+public interface ConfigBuilder {
+
+    String build(boolean isEn);
+}
