+++ datavines-connector/datavines-connector-plugins/datavines-connector-jdbc/src/main/java/io/datavines/connector/plugin/JdbcConnector.java
@@ -16,21 +16,12 @@
  */
 package io.datavines.connector.plugin;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import io.datavines.common.CommonConstants;
 import io.datavines.common.datasource.jdbc.*;
 import io.datavines.common.datasource.jdbc.entity.ColumnInfo;
 import io.datavines.common.datasource.jdbc.entity.DatabaseInfo;
 import io.datavines.common.datasource.jdbc.entity.TableInfo;
 import io.datavines.common.datasource.jdbc.entity.TableColumnInfo;
 import io.datavines.common.param.*;
-import io.datavines.common.param.form.PluginParams;
-import io.datavines.common.param.form.PropsType;
-import io.datavines.common.param.form.Validate;
-import io.datavines.common.param.form.props.InputParamsProps;
-import io.datavines.common.param.form.type.InputParam;
 import io.datavines.common.utils.JSONUtils;
 import io.datavines.connector.api.Connector;
 import io.datavines.common.datasource.jdbc.utils.JdbcDataSourceUtils;
@@ -198,32 +189,6 @@ public abstract class JdbcConnector implements Connector, IJdbcDataSourceInfo {
         return ConnectorResponse.builder().status(ConnectorResponse.Status.SUCCESS).result(false).build();
     }
 
-    @Override
-    public String getConfigJson(boolean isEn) {
-        List<PluginParams> params = new ArrayList<>();
-        params.add(getHostInput(isEn));
-        params.add(getPortInput(isEn));
-        if (getCatalogInput(isEn) != null) {
-            params.add(getCatalogInput(isEn));
-        }
-        params.add(getDatabaseInput(isEn));
-        params.add(getUserInput(isEn));
-        params.add(getPasswordInput(isEn));
-        params.add(getPropertiesInput(isEn));
-
-        ObjectMapper mapper = new ObjectMapper();
-        mapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
-        String result = null;
-
-        try {
-            result = mapper.writeValueAsString(params);
-        } catch (JsonProcessingException e) {
-            logger.error("json parse error : ", e);
-        }
-
-        return result;
-    }
-
     private List<String> getPrimaryKeys(String catalog, String schema, String tableName, DatabaseMetaData metaData) {
         ResultSet rs = null;
         List<String> primaryKeys = new ArrayList<>();
@@ -290,70 +255,4 @@ public abstract class JdbcConnector implements Connector, IJdbcDataSourceInfo {
     protected ResultSet getPrimaryKeys(DatabaseMetaData metaData,String catalog, String schema, String tableName) throws SQLException {
         return metaData.getPrimaryKeys(schema, null, tableName);
     }
-
-    protected InputParam getHostInput(boolean isEn) {
-        return getInputParam("host",
-                isEn ? "host":"地址",
-                isEn ? "please enter host ip" : "请填入连接地址", 1,
-                Validate.newBuilder().setRequired(true).setMessage(isEn ? "please enter host ip" : "请填入连接地址").build(),
-                null);
-    }
-
-    protected InputParam getPortInput(boolean isEn) {
-        return getInputParam("port",
-                isEn ? "port" : "端口",
-                isEn ? "please enter port" : "请填入端口号", 1,
-                Validate.newBuilder().setRequired(true).setMessage(isEn ? "please enter port" : "请填入端口号").build(),
-                null);
-    }
-
-    protected InputParam getCatalogInput(boolean isEn) {
-        return null;
-    }
-
-    protected InputParam getDatabaseInput(boolean isEn) {
-        return getInputParam("database",
-                isEn ? "database" : "数据库",
-                isEn ? "please enter database" : "请填入数据库", 1,
-                Validate.newBuilder().setRequired(true).setMessage(isEn ? "please enter database" : "请填入数据库").build(),
-                null);
-    }
-
-    protected InputParam getUserInput(boolean isEn) {
-        return getInputParam("user",
-                isEn ? "user" : "用户名",
-                isEn ? "please enter user" : "请填入用户名", 1,
-                Validate.newBuilder().setRequired(true).setMessage(isEn ? "please enter user" : "请填入用户名").build(),
-                null);
-    }
-
-    protected InputParam getPasswordInput(boolean isEn) {
-        return getInputParam("password",
-                isEn ? "password" : "密码",
-                isEn ? "please enter password" : "请填入密码", 1,
-                Validate.newBuilder().setRequired(true).setMessage(isEn ? "please enter password" : "请填入密码").build(),
-                null);
-    }
-
-    protected InputParam getPropertiesInput(boolean isEn) {
-        return getInputParam("properties",
-                isEn ? "properties" : "参数",
-                isEn ? "please enter properties,like key=value&key1=value1" : "请填入参数，格式为key=value&key1=value1", 2, null,
-                null);
-    }
-
-    protected InputParam getInputParam(String field, String title, String placeholder, int rows, Validate validate , Object defaultValue) {
-        return InputParam
-                .newBuilder(field, title)
-                .addValidate(validate)
-                .setProps(new InputParamsProps().setDisabled(false))
-                .setSize(CommonConstants.SMALL)
-                .setType(PropsType.TEXT)
-                .setRows(rows)
-                .setPlaceholder(placeholder)
-                .setValue(defaultValue)
-                .setEmit(null)
-                .build();
-    }
-
 }
