+++ datavines-connector/datavines-connector-plugins/datavines-connector-mysql/src/main/java/io/datavines/connector/plugin/MysqlConnector.java
@@ -48,11 +48,4 @@ public class MysqlConnector extends JdbcConnector {
         return metaData.getCatalogs();
     }
 
-    @Override
-    protected InputParam getPropertiesInput(boolean isEn) {
-        return getInputParam("properties",
-                isEn ? "properties" : "参数",
-                isEn ? "please enter properties,like key=value&key1=value1" : "请填入参数，格式为key=value&key1=value1", 2, null,
-                "useUnicode=true&characterEncoding=UTF-8&useSSL=false&serverTimezone=Asia/Shanghai");
-    }
 }
