+++ datavines-connector/datavines-connector-plugins/datavines-connector-presto/src/main/java/io/datavines/connector/plugin/PrestoConnectorFactory.java
@@ -16,10 +16,7 @@
  */
 package io.datavines.connector.plugin;
 
-import io.datavines.connector.api.Connector;
-import io.datavines.connector.api.ConnectorParameterConverter;
-import io.datavines.connector.api.Dialect;
-import io.datavines.connector.api.Executor;
+import io.datavines.connector.api.*;
 
 public class PrestoConnectorFactory extends AbstractJdbcConnectorFactory {
 
@@ -41,5 +38,10 @@ public class PrestoConnectorFactory extends AbstractJdbcConnectorFactory {
     @Override
     public Executor getExecutor() {
         return new PrestoExecutor();
+    }
+
+    @Override
+    public ConfigBuilder getConfigBuilder() {
+        return new PrestoConfigBuilder();
     }
 }
