+++ datavines-server/src/main/resources/application.yaml
@@ -58,6 +58,8 @@ spring:
 
 mybatis-plus:
   type-enums-package: io.datavines.*.enums
+  configuration:
+    log-impl: org.apache.ibatis.logging.stdout.StdOutImpl # 打印日志到控制台
 
 server:
   port: 5600
